feat: initial commit
A nodejs object which provides data for the game Apex Legends. Current
scope is seasonal information (name, dates, maps, playlist rotations)
and basic information about characters (name, class, tagline).

This was spun off from a selection of classes I was using in a different
project. The functionality may become useful as the basis for an API in
the near future, so I've split it into a separate module for easier
reuse and to help keep my projects a bit cleaner.