refactor!: remove queryDate property from classes
queryDate doesn't feel like it belongs as a property - queries to dates
should be handled using methods. `currentPlaylistTimeElapsed()` should
still be able to be provided a date as an optional argument so it has
also been moved to a method.

BREAKING CHANGE: removed `queryDate` property, moved
`currentPlaylistTimeElapsed()` to a get~ method.