feat!: add mapByDate and scheduled items
Added a feature to query a map based on a date. It accepts a date as an
argument, but it uses the current date if none is provided, which makes
it an easy replacement for the .currentMap getter too. I have had to
change several other methods for this to work. The object it returns
extends PlaylistItem with a start and end time, and a getter for the
remaining time. The overall interface is much nicer, it means I no
longer have to fetch the next map item to calculate the remaining time,
but it did mean rewriting both classes quite extensively. I'm now also
using a date mocker for tests rather than passing in dates as arguments,
which is much neater.

BREAKING CHANGE: arguments for PlaylistItem