feat(data): update season 11 and 12 data
Added new data for season 12 as it's coming in. Seasons appear to start
at 1800 GMT, so I have corrected the season 11 start date. However, no
order of map durations seems to fall at 1800, so I will need to add a
new property set to midday on the Playlist otherwise the results will be
offset.