feat(Playlist): add rotationBaseTime getter
This is needed to correct for the difference between when the playlist
rotation appears to begin and the official start date of the season. I
don't know why the map rotations can't be fit squarely on the start of
the season start. Perhaps Respawn start ramping them up early for
streamers. Maybe it's different every season.