feat: add getMapsByDate() to ApexObj and Season
New method which gets an array of the maps currently playing for each
playlist. This is aliased at the root level for easy searching. At the
moment the module only supports the default playlists, not takeover
events etc, so both it will need refactoring later once I've decided how
to implement takeover events. But it passes the tests for now, so it can
stay like this for now.