feat(Season): add parsePlaylist method
This method is a factory function which parses information from the
supplied data to return a Playlist instance of the correct type,
currently either SplitPlaylist or RotatingPlaylist. This is to abstract
away the logic from the constructor and to support the upcoming addition
of LTMs, which will also necessitate a new Playlist subclass for
single-map playlists.