refactor!: change season mode names
I have been using the colloquial names my regular squad uses for each of
the modes, but to keep things in sync with 'the Respawn way' it is time
to switch to using the official mode names. I resisted this initially
because 'Play Apex' is such an awkward mode name imho compared to the
pleasing symmetry of [ranked/unranked][battleRoyale/arenas].

However, there are whispers that Respawn are looking at ways to branch
out into modes other than BR, such as with the new Control event in
season 12. The types of modes and the number of ranked variants might
not be symmetrical. BR might not always been an option, or at least not
a given. Arenas might split mid-season with Control rather than the
current rank reset (but no other changes, i.e. same maps). Better to
keep the modes looser and filter explicitly.

BREAKING CHANGE: mode names in JSON source