refactor(RotatingPlaylist)!: remove .mapDurations
This property only appears on the RotatingPlaylist variant of Playlist,
and isn't referenced by anything outside of the class definition, so
there is no need for it to be a public property. I have removed it and
the associated test.

As I'm messing around with mapDurations I also took this opportunity to
refactor `this.rotations` to use array destructuring rather than
creating new arrays via `Array.from()`. The latter method has a built-in
`.map` function as a second argument and (I suspect) is negligabally
more performant, but I find the destructuring form easier to follow at a
glance.

BREAKING CHANGE: remove .mapDurations property from RotatingPlaylist