refactor(ScheduledPlaylistItem)!: modify endTime
I have stopped removing a millisecond from the end of
`ScheduledPlaylistItem.endTime` as the data is imported from JSON. This
seemed to make sense at the time because it meant no maps would overlap.
However I have decided it would be more useful to have the `endTime` of
one `ScheduledPlaylistItem` be equal to the start of the next; this
makes it super easy to query the next item using the `endTime` of the
current one.

BREAKING CHANGE: `ScheduledPlaylistItem.endTime` no longer -1ms