feat(Season): add support for LTMs
Apex occasionally has limited time modes (LTMs), and these are
increasing in frequency over time it appears. There are two at the start
of season 12: a new mode called Control which is in addition to the
regular playlists, and an un-named event (as far as I can see) at the
start of each season where the newly-modified map is the only one
available in the standard 'Play Apex' mode for the first week (I have
called this Olympus 24/7 to follow the convention used when Respawn was
Infinity Ward and working on Call of Duty, and continued into
Titanfall).

I played around with a couple of different ways of implementing this. I
think the most flexible is including them in the `playlists` property on
Season, and adding the properties `LTM`, `takeover` and `replaces` to
PlaylistItem. I experimented with adding LTMs as a separate property,
but actually I think it makes more logical sense to treat them like any
other playlist and allow the consumer to filter them by the new
PlaylistItem properties as needed. This also greatly simplifies
filtering them out of the `getMapsByDate(date)` method and `currentMaps`
computed property.

They are still declared in a separate `LTMs` property in the JSON
source, however. My goal is to keep the JSON as clear as possible to
keep updates trivial, and having a separate section to delcare new LTMs
as they are announced supports this.