fix(ApexObj): correct getSeasonByDate startTime
The `getSeasonByDate(date)` method was checking the `startTime` of each
season using 'less-than' rather than 'less-than-or-equal-to', meaning it
wouldn't fetch seasons at the exact start date. Dates are accurate to
the millisecond so this would almost certainly never crop up in
production, but it still needs to be fixed because it allows testing to
search using the exact start time of a particular season for
convenience.