refactor(util): fold isParseableDate -> parseDate
I was writing a lot of boilerplate code to check dates when I realised I
can just have `parseDate(date)` to return the current time as a Date
object if the `date` argument is missing or throw an error if the
argument isn't parseable. This saves 4-5 lines of boilerplate code
almost everywhere dates are passed as arguments.