feat(ApexObj): add nextMaps pseudo property
ApexObj will now alias the nextMaps pseudo property of the current
season. At some point before v1.0.0 this will need to be refactored to
check if more maps are available in the next season, e.g. during the
next Ranked Leagues split after the maps for the next season of Ranked
Leagues are announced. This is some time off however (the new maps
aren't usually announced until a couple weeks before the next season)
and this functionality passes the tests for now.

I have more functionality planned that will make the above easier (a
getPlaylistsByDate method for Season, and references to the current
Playlist and Season for PlaylistItem), so there's no point writing a
bunch of code for nextMaps only to have to rewrite two versions later
before the second Ranked Leagues split has even begun.