feat(util): add withinDates function
Doing a lot of checking whether a given class instance is within certain
date ranges, lot of boilerplate code. Every class has a `startTime` and
`endTime` so straightforward to create a utility function to check all
of these by passing `this` as the first function and the target date as
the second.