feat(Playlist)!: add Placelist.baseTime property
Added a baseTime property to the json source data and to the base
Playlist class for playlists which require a startTime override for
their rotation logic to sync with the real-world map rotations. The map
rotations don't appear to start at either the season start time or GMT,
and appear inconsistent between seasons. As technically pleasing as it
would be to automatically find when the rotations start, for now just
hard-coding it in the source data allows for much easier patching.

BREAKING CHANGE: rotating playlists require a baseTime