chore(package): update description and keywords
Improved keywords to aid searching in npm registry, and added extra
information to the description. Particularly highlighting that the
module does not depend on the mozambiquehe.re API, unlike almost all
other Apex Legends packages - this is a major differentiator for this
package, and one of the reasons it exists is that the aforementioned API
is constantly going offline.